Orchestrator Refactoring
Refactored orchestrator so that the two separate concerns of ingestion and querying were separated into their own classes. Also fixed a chunking bug.

Closes #62
