+++ Generellem.Tests/IngestionTests.cs
@@ -2,36 +2,31 @@
 
 using Generellem.Document.DocumentTypes;
 using Generellem.DocumentSource;
-using Generellem.Llm;
-using Generellem.Llm.AzureOpenAI;
 using Generellem.Rag;
 using Generellem.Repository;
 using Generellem.Services;
 
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace Generellem.Orchestrator.Tests;
+namespace Generellem.Processors.Tests;
 
-public class AzureOpenAIOrchestratorTests
+public class IngestionTests
 {
     readonly string DocSource = $"{Environment.MachineName}:{nameof(FileSystem)}";
 
-    readonly Mock<IConfiguration> configMock = new();
     readonly Mock<IDocumentHashRepository> docHashRepMock = new();
     readonly Mock<IDocumentSource> docSourceMock = new();
     readonly Mock<IDocumentSourceFactory> docSourceFactoryMock = new();
     readonly Mock<IDocumentType> docTypeMock = new();
-    readonly Mock<ILlm> llmMock = new();
-    readonly Mock<ILogger<AzureOpenAIOrchestrator>> logMock = new();
+    readonly Mock<ILogger<Ingestion>> logMock = new();
     readonly Mock<IRag> ragMock = new();
 
-    readonly AzureOpenAIOrchestrator orchestrator;
+    readonly Ingestion ingestion;
     readonly Queue<ChatMessage> chatHistory = new();
 
     readonly List<IDocumentSource> docSources = [];
 
-    public AzureOpenAIOrchestratorTests()
+    public IngestionTests()
     {
         docSources.Add(docSourceMock.Object);
 
@@ -41,97 +36,8 @@ public class AzureOpenAIOrchestratorTests
         ragMock
             .Setup(rag => rag.SearchAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
             .Returns(Task.FromResult(new List<string> { "Search Result" }));
-        llmMock
-            .Setup(llm => llm.AskAsync<AzureOpenAIChatResponse>(It.IsAny<IChatRequest>(), It.IsAny<CancellationToken>()))
-            .Returns(Task.FromResult(Mock.Of<AzureOpenAIChatResponse>()));
 
-        orchestrator = new AzureOpenAIOrchestrator(
-            configMock.Object, docHashRepMock.Object, docSourceFactoryMock.Object, llmMock.Object, logMock.Object, ragMock.Object);
-    }
-
-    [Fact]
-    public async Task AskAsync_WithNullDeploymentName_ThrowsArgumentNullException()
-    {
-        configMock.SetupGet(config => config[GKeys.AzOpenAIDeploymentName]).Returns(value: null);
-
-        await Assert.ThrowsAsync<ArgumentNullException>(() =>
-          orchestrator.AskAsync("Hello", chatHistory, CancellationToken.None));
-    }
-
-    [Fact]
-    public async Task AskAsync_WithEmptyDeploymentName_ThrowsArgumentNullException()
-    {
-        configMock.Setup(config => config[GKeys.AzOpenAIDeploymentName]).Returns(string.Empty);
-
-        await Assert.ThrowsAsync<ArgumentException>(() =>
-            orchestrator.AskAsync("Hello", chatHistory, CancellationToken.None));
-    }
-
-    [Fact]
-    public async Task AskAsync_CallsSearchAsync()
-    {
-        Mock<IConfigurationSection> configSection = new();
-
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIDeploymentName])
-            .Returns("generellem");
-
-        await orchestrator.AskAsync("Hello", chatHistory, CancellationToken.None);
-
-        ragMock.Verify(rag => rag.SearchAsync(It.IsAny<string>(), CancellationToken.None), Times.Once());
-    }
-
-    [Fact]
-    public async Task AskAsync_CallsAskAsync()
-    {
-        Mock<IConfigurationSection> configSection = new();
-
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIDeploymentName])
-            .Returns("generellem");
-
-        await orchestrator.AskAsync("Hello", chatHistory, CancellationToken.None);
-
-        llmMock.Verify(llm => llm.AskAsync<AzureOpenAIChatResponse>(It.IsAny<IChatRequest>(), It.IsAny<CancellationToken>()), Times.Exactly(2));
-    }
-
-    [Fact]
-    public async Task AskAsync_BuildsContextStringFromSearchResults()
-    {
-        var searchResults = new List<string> { "result1", "result2" };
-        ragMock
-            .Setup(rag => rag.SearchAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(searchResults);
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIDeploymentName])
-            .Returns("generellem");
-
-        await orchestrator.AskAsync("Hello", chatHistory, CancellationToken.None);
-
-        llmMock.Verify(llm =>
-            llm.AskAsync<AzureOpenAIChatResponse>(
-                It.IsAny<AzureOpenAIChatRequest>(),
-                It.IsAny<CancellationToken>()));
-    }
-
-    [Fact]
-    public async Task AskAsync_PopulatesChatHistory()
-    {
-        const string ExpectedQuery = "What is Generellem?";
-        var searchResults = new List<string> { "result1", "result2" };
-        ragMock
-            .Setup(rag => rag.SearchAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(searchResults);
-        configMock
-            .Setup(config => config[GKeys.AzOpenAIDeploymentName])
-            .Returns("generellem");
-
-        await orchestrator.AskAsync(ExpectedQuery, chatHistory, CancellationToken.None);
-
-        Assert.Single(chatHistory);
-        ChatMessage chatMessage = chatHistory.Peek();
-        Assert.Equal(ChatRole.User, chatMessage.Role);
-        Assert.Equal(ExpectedQuery, chatMessage.Content);
+        ingestion = new Ingestion(docHashRepMock.Object, docSourceFactoryMock.Object, logMock.Object, ragMock.Object);
     }
 
     void SetupGetDocumentsAsync(string filePath)
@@ -149,7 +55,7 @@ public class AzureOpenAIOrchestratorTests
     {
         SetupGetDocumentsAsync("TestDocs\\file.txt");
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         docSourceMock.Verify(docSrc => docSrc.GetDocumentsAsync(It.IsAny<CancellationToken>()), Times.Once());
     }
@@ -168,7 +74,7 @@ public class AzureOpenAIOrchestratorTests
         }
         docSourceMock.Setup(docSrc => docSrc.GetDocumentsAsync(It.IsAny<CancellationToken>())).Returns(GetDocInfos);
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         docTypeMock.Verify(
             docType => docType.GetTextAsync(It.IsAny<Stream>(), It.IsAny<string>()),
@@ -189,7 +95,7 @@ public class AzureOpenAIOrchestratorTests
         }
         docSourceMock.Setup(docSrc => docSrc.GetDocumentsAsync(It.IsAny<CancellationToken>())).Returns(GetDocInfos);
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         logMock
             .Verify(
@@ -207,7 +113,7 @@ public class AzureOpenAIOrchestratorTests
     {
         SetupGetDocumentsAsync("TestDocs\\file.txt");
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         ragMock.Verify(
             rag => rag.EmbedAsync(It.IsAny<string>(), It.IsAny<IDocumentType>(), It.IsAny<string>(), CancellationToken.None), 
@@ -225,7 +131,7 @@ public class AzureOpenAIOrchestratorTests
         }
         docSourceMock.Setup(docSrc => docSrc.GetDocumentsAsync(It.IsAny<CancellationToken>())).Returns(GetDocInfos);
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         ragMock.Verify(
             rag => rag.EmbedAsync(It.IsAny<string>(), It.IsAny<IDocumentType>(), It.IsAny<string>(), CancellationToken.None),
@@ -240,7 +146,7 @@ public class AzureOpenAIOrchestratorTests
             .Setup(docHashRep => docHashRep.GetDocumentHash(It.IsAny<string>()))
             .Returns((DocumentHash?)null);
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         docHashRepMock.Verify(
             docHashRep => docHashRep.Insert(It.IsAny<DocumentHash>()),
@@ -255,7 +161,7 @@ public class AzureOpenAIOrchestratorTests
             .Setup(docHashRep => docHashRep.GetDocumentHash(It.IsAny<string>()))
             .Returns(new DocumentHash { DocumentReference = "", Hash = Guid.NewGuid().ToString() });
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         docHashRepMock.Verify(
             docHashRep => docHashRep.Update(It.IsAny<DocumentHash>(), It.IsAny<string>()),
@@ -273,7 +179,7 @@ public class AzureOpenAIOrchestratorTests
             .Setup(docHashRep => docHashRep.GetDocumentHash(It.IsAny<string>()))
             .Returns(new DocumentHash { DocumentReference = "", Hash = SHA256BlankStringHash });
 
-        await orchestrator.ProcessFilesAsync(CancellationToken.None);
+        await ingestion.IngestDocumentsAsync(CancellationToken.None);
 
         docHashRepMock.Verify(
             docHashRep => docHashRep.Update(It.IsAny<DocumentHash>(), It.IsAny<string>()),
